+++ test/integration/security/access_sanity.js
@@ -1,23 +1,37 @@
 describe(
   require('../../__fixtures/utils/test_helper')
-    .create()
-    .testName(__filename, 3),
+  .create()
+  .testName(__filename, 3),
   function() {
-    var happn = require('../../../lib/index');
-    var serviceInstance;
-    var adminClient;
-    var expect = require('expect.js');
-    var test_id = Date.now() + '_' + require('shortid').generate();
+    const request = require('request');
+    const happn = require('../../../lib/index');
+    let serviceInstance;
+    let adminClient;
+    const expect = require('expect.js');
+    const test_id = Date.now() + '_' + require('shortid').generate();
+    let testClient, testClient1;
 
-    var testClient;
-
-    var getService = function(config, callback) {
+    function getService(config, callback) {
       happn.service.create(config, callback);
-    };
+    }
+
+    function doRequest(path, token) {
+      return new Promise((resolve, reject) => {
+        let options = {
+          url: 'http://127.0.0.1:55000' + path
+        };
+        options.headers = {
+          Cookie: ['happn_token=' + token]
+        };
+        request(options, function(error, response) {
+          if (error) return reject(error);
+          resolve(response);
+        });
+      });
+    }
 
     before('it starts completely defaulted service', function(done) {
-      getService(
-        {
+      getService({
           secure: true
         },
         function(e, service) {
@@ -159,6 +173,11 @@ describe(
       ] = {
         actions: ['set']
       };
+      testGroup.permissions[
+        '/@HTTP/TEST/a7_eventemitter_security_access/' + test_id + '/web_access'
+      ] = {
+        actions: ['*']
+      };
 
       var testUser = {
         username: 'TEST USER@blah.com' + test_id,
@@ -168,56 +187,75 @@ describe(
         }
       };
 
-      var addedTestGroup;
-      var addedTestuser;
+      var testUser1 = {
+        username: 'TEST USE1R@blah.com' + test_id,
+        password: 'TEST PWD',
+        custom_data: {
+          something: 'usefull'
+        }
+      };
+
+      let addedTestGroup;
+      let addedTestuser;
+      var addedTestuser1;
+
+      function recreateAddedTestUser(){
+        return new Promise(async (resolve, reject) => {
+          try{
+            addedTestuser1 = await serviceInstance.services.security.users.upsertUser(
+              testUser1, {
+                overwrite: false
+              }
+            );
+            await serviceInstance.services.security.users.linkGroup(
+              addedTestGroup,
+              addedTestuser1);
+            resolve();
+          }catch(e){
+            reject(e);
+          }
+        });
+      }
 
       before(
         'creates a group and a user, adds the group to the user, logs in with test user',
-        function(done) {
-          serviceInstance.services.security.users.upsertGroup(
-            testGroup,
-            {
+        async () => {
+          addedTestGroup = await serviceInstance.services.security.users.upsertGroup(
+            testGroup, {
               overwrite: false
-            },
-            function(e, result) {
-              if (e) return done(e);
-              addedTestGroup = result;
-
-              serviceInstance.services.security.users.upsertUser(
-                testUser,
-                {
-                  overwrite: false
-                },
-                function(e, result) {
-                  if (e) return done(e);
-                  addedTestuser = result;
-
-                  serviceInstance.services.security.users.linkGroup(
-                    addedTestGroup,
-                    addedTestuser,
-                    function(e) {
-                      if (e) return done(e);
-
-                      serviceInstance.services.session
-                        .localClient({
-                          username: testUser.username,
-                          password: 'TEST PWD'
-                        })
-
-                        .then(function(clientInstance) {
-                          testClient = clientInstance;
-                          done();
-                        })
-
-                        .catch(function(e) {
-                          done(e);
-                        });
-                    }
-                  );
-                }
-              );
             }
           );
+          addedTestuser = await serviceInstance.services.security.users.upsertUser(
+            testUser, {
+              overwrite: false
+            }
+          );
+
+          addedTestuser1 = await serviceInstance.services.security.users.upsertUser(
+            testUser1, {
+              overwrite: false
+            }
+          );
+
+          await serviceInstance.services.security.users.linkGroup(
+            addedTestGroup,
+            addedTestuser);
+
+          await serviceInstance.services.security.users.linkGroup(
+            addedTestGroup,
+            addedTestuser1);
+
+          testClient = await serviceInstance.services.session
+            .localClient({
+              username: testUser.username,
+              password: 'TEST PWD'
+            });
+
+          testClient1 = await serviceInstance.services.session
+            .localClient({
+              username: testUser1.username,
+              password: 'TEST PWD'
+            });
         }
       );
 
@@ -246,15 +284,13 @@ describe(
 
       it('checks allowed on, and prevented from on', function(done) {
         testClient.on(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/on',
-          {},
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/on', {},
           function() {},
           function(e) {
             if (e) return done(e);
 
             testClient.on(
-              '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/on',
-              {},
+              '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/on', {},
               function() {},
               function(e) {
                 if (!e)
@@ -274,8 +310,7 @@ describe(
 
       it('delegated authority: checks allowed on, and prevented from on', function(done) {
         adminClient.on(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/on', {
             onBehalfOf: testClient.session.user.username
           },
           function() {},
@@ -283,8 +318,7 @@ describe(
             if (e) return done(e);
 
             adminClient.on(
-              '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/on',
-              {
+              '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/on', {
                 onBehalfOf: testClient.session.user.username
               },
               function() {},
@@ -316,11 +350,9 @@ describe(
           );
 
           testClient.set(
-            '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/set',
-            {
+            '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/set', {
               test: 'test'
-            },
-            {},
+            }, {},
             function(e) {
               if (!e)
                 return done(
@@ -335,9 +367,7 @@ describe(
 
       it('delegated authority: checks allowed set, and prevented from set', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/set',
-          {},
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/set', {}, {
             onBehalfOf: testClient.session.user.username
           },
           function(e, result) {
@@ -348,11 +378,9 @@ describe(
             );
 
             adminClient.set(
-              '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/set',
-              {
+              '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/set', {
                 test: 'test'
-              },
-              {
+              }, {
                 onBehalfOf: testClient.session.user.username
               },
               function(e) {
@@ -370,8 +398,7 @@ describe(
 
       it('delegated authority: checks prevented from getPaths', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/getPaths/1',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/getPaths/1', {
             property1: 'property1',
             property2: 'property2',
             property3: 'property3'
@@ -380,13 +407,11 @@ describe(
           function(e) {
             expect(e == null).to.be(true);
             adminClient.set(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/getPaths/2',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/getPaths/2', {
                 property1: 'property1',
                 property2: 'property2',
                 property3: 'property3'
-              },
-              {
+              }, {
                 onBehalfOf: testClient.session.user.username
               },
               function(e) {
@@ -406,8 +431,7 @@ describe(
         adminClient.increment(
           '/TEST/a7_eventemitter_security_access/' + test_id + '/increment',
           'counter',
-          1,
-          {
+          1, {
             onBehalfOf: testClient.session.user.username
           },
           function(e) {
@@ -421,9 +445,9 @@ describe(
 
       it('delegated authority: checks prevented from setSibling', function(done) {
         adminClient.setSibling(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/setSibling',
-          { property1: 'property1' },
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/setSibling', {
+            property1: 'property1'
+          }, {
             onBehalfOf: testClient.session.user.username
           },
           function(e) {
@@ -437,17 +461,14 @@ describe(
 
       it('checks allowed get, and prevented from get', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
 
             testClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {},
               function(e, result) {
                 if (e) return done(e);
                 expect(result._meta.path).to.be(
@@ -455,8 +476,7 @@ describe(
                 );
 
                 testClient.get(
-                  '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/get',
-                  {},
+                  '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/get', {},
                   function(e) {
                     if (!e)
                       return done(
@@ -474,17 +494,14 @@ describe(
 
       it('delegated authority: checks allowed get, and prevented from get', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
 
             adminClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {
                 onBehalfOf: testClient.session.user.username
               },
               function(e, result) {
@@ -494,8 +511,7 @@ describe(
                 );
 
                 adminClient.get(
-                  '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/get',
-                  {
+                  '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/get', {
                     onBehalfOf: testClient.session.user.username
                   },
                   function(e) {
@@ -515,24 +531,20 @@ describe(
 
       it('checks allowed count, and prevented from count', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/count',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/count', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
 
             testClient.count(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/count',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/count', {},
               function(e, result) {
                 if (e) return done(e);
                 expect(result.value).to.be(1);
 
                 testClient.get(
-                  '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/count',
-                  {},
+                  '/TEST/a7_eventemitter_security_access/dodge/' + test_id + '/count', {},
                   function(e) {
                     if (!e)
                       return done(
@@ -559,11 +571,9 @@ describe(
           );
 
           testClient.set(
-            '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-            {
+            '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {
               test: 'test'
-            },
-            {},
+            }, {},
             function(e) {
               if (!e)
                 return done(
@@ -578,8 +588,7 @@ describe(
 
       it('delegated authority: checks allowed get but not set', function(done) {
         adminClient.get(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {
             onBehalfOf: testClient.session.user.username
           },
           function(e, result) {
@@ -589,11 +598,9 @@ describe(
             );
 
             adminClient.set(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/get',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/get', {
                 test: 'test'
-              },
-              {
+              }, {
                 onBehalfOf: testClient.session.user.username
               },
               function(e) {
@@ -611,16 +618,13 @@ describe(
 
       it('checks allowed get and on but not set', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             testClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {},
               function(e, result) {
                 if (e) return done(e);
                 expect(result._meta.path).to.be(
@@ -628,18 +632,15 @@ describe(
                 );
 
                 testClient.on(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-                  {},
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {},
                   function() {},
                   function(e) {
                     if (e) return done(e);
 
                     testClient.set(
-                      '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-                      {
+                      '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {
                         test: 'test'
-                      },
-                      {},
+                      }, {},
                       function(e) {
                         if (!e)
                           return done(
@@ -659,16 +660,13 @@ describe(
 
       it('delegated authority: checks allowed get and on but not set', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             adminClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {
                 onBehalfOf: testClient.session.user.username
               },
               function(e, result) {
@@ -678,8 +676,7 @@ describe(
                 );
 
                 adminClient.on(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-                  {
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {
                     onBehalfOf: testClient.session.user.username
                   },
                   function() {},
@@ -687,11 +684,9 @@ describe(
                     if (e) return done(e);
 
                     adminClient.set(
-                      '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on',
-                      {
+                      '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_on', {
                         test: 'test'
-                      },
-                      {
+                      }, {
                         onBehalfOf: testClient.session.user.username
                       },
                       function(e) {
@@ -713,16 +708,13 @@ describe(
 
       it('checks allowed get but not on', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             testClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on', {},
               function(e, result) {
                 if (e) return done(e);
                 expect(result._meta.path).to.be(
@@ -730,8 +722,7 @@ describe(
                 );
 
                 testClient.on(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on',
-                  {},
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on', {},
                   function() {},
                   function(e) {
                     if (!e) return done(new Error('this should not have been allowed...'));
@@ -747,16 +738,13 @@ describe(
 
       it('delegated authority: checks allowed get but not on', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             adminClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on', {
                 onBehalfOf: testClient.session.user.username
               },
               function(e, result) {
@@ -766,8 +754,7 @@ describe(
                 );
 
                 adminClient.on(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on',
-                  {
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/get_not_on', {
                     onBehalfOf: testClient.session.user.username
                   },
                   function() {},
@@ -785,23 +772,19 @@ describe(
 
       it('checks allowed on but not get', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             testClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get', {},
               function(e) {
                 if (!e) return done(new Error('this should not have been allowed...'));
                 expect(e.toString()).to.be('AccessDenied: unauthorized');
 
                 testClient.on(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get',
-                  {},
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get', {},
                   function() {},
                   done
                 );
@@ -813,16 +796,13 @@ describe(
 
       it('delegated authority: checks allowed on but not get', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             adminClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get', {
                 onBehalfOf: testClient.session.user.username
               },
               function(e) {
@@ -830,8 +810,7 @@ describe(
                 expect(e.toString()).to.be('AccessDenied: unauthorized');
 
                 adminClient.on(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get',
-                  {
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/on_not_get', {
                     onBehalfOf: testClient.session.user.username
                   },
                   function() {},
@@ -845,16 +824,13 @@ describe(
 
       it('checks allowed set but not get', function(done) {
         testClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             testClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get', {},
               function(e) {
                 if (!e) return done(new Error('this should not have been allowed...'));
                 expect(e.toString()).to.be('AccessDenied: unauthorized');
@@ -867,18 +843,15 @@ describe(
 
       it('delegated authority: checks allowed set but not get', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get', {
             'test-set': 'test-set-val'
-          },
-          {
+          }, {
             onBehalfOf: testClient.session.user.username
           },
           function(e) {
             if (e) return done(e);
             adminClient.get(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_get', {
                 onBehalfOf: testClient.session.user.username
               },
               function(e) {
@@ -893,16 +866,13 @@ describe(
 
       it('checks allowed set but not on', function(done) {
         testClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on', {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           function(e) {
             if (e) return done(e);
             testClient.on(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on',
-              {},
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on', {},
               function() {},
               function(e) {
                 if (!e) return done(new Error('this should not have been allowed...'));
@@ -916,18 +886,15 @@ describe(
 
       it('delegated authority: checks allowed set but not on', function(done) {
         adminClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on', {
             'test-set': 'test-set-val'
-          },
-          {
+          }, {
             onBehalfOf: testClient.session.user.username
           },
           function(e) {
             if (e) return done(e);
             testClient.on(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/comp/set_not_on', {
                 onBehalfOf: testClient.session.user.username
               },
               function() {},
@@ -943,16 +910,14 @@ describe(
 
       it('checks allowed get all', function(done) {
         testClient.get(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/get_all/' + test_id,
-          {},
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/get_all/' + test_id, {},
           done
         );
       });
 
       it('checks allowed on all', function(done) {
         testClient.on(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/on_all/' + test_id,
-          {},
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/on_all/' + test_id, {},
           function() {},
           done
         );
@@ -960,19 +925,16 @@ describe(
 
       it('checks allowed set all', function(done) {
         testClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/set_all/' + test_id,
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/set_all/' + test_id, {
             'test-set': 'test-set-val'
-          },
-          {},
+          }, {},
           done
         );
       });
 
       it('checks against a permission that doesnt exist', function(done) {
         testClient.get(
-          '/TEST/a7_eventemitter_security_access/whatevs' + test_id + '/get_all/' + test_id,
-          {},
+          '/TEST/a7_eventemitter_security_access/whatevs' + test_id + '/get_all/' + test_id, {},
           function(e) {
             if (!e) return done(new Error('this should not have been allowed...'));
             expect(e.toString()).to.be('AccessDenied: unauthorized');
@@ -991,11 +953,9 @@ describe(
 
           setTimeout(function() {
             testClient.set(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/set',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/set', {
                 test: 'data'
-              },
-              {},
+              }, {},
               function(e) {
                 if (!e) return done(new Error('this should not have been allowed...'));
                 expect(e.toString()).to.be('AccessDenied: unauthorized');
@@ -1016,11 +976,9 @@ describe(
 
           setTimeout(function() {
             testClient.set(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/set',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/set', {
                 test: 'data'
-              },
-              {},
+              }, {},
               done
             );
           }, 2000);
@@ -1029,8 +987,7 @@ describe(
 
       it('tests the remove permission', function(done) {
         testClient.set(
-          '/TEST/a7_eventemitter_security_access/' + test_id + '/remove-permission',
-          {
+          '/TEST/a7_eventemitter_security_access/' + test_id + '/remove-permission', {
             test: 'data'
           },
           function(e) {
@@ -1045,8 +1002,7 @@ describe(
               )
               .then(function() {
                 testClient.set(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/remove-permission',
-                  {
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/remove-permission', {
                     test: 'data'
                   },
                   function(e) {
@@ -1069,8 +1025,7 @@ describe(
           )
           .then(function() {
             testClient.set(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/remove-permission',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/remove-permission', {
                 test: 'data'
               },
               done
@@ -1084,8 +1039,7 @@ describe(
           '/TEST/a7_eventemitter_security_access/' + test_id + '/prohibit-permission';
 
         testClient.set(
-          prohibitPath,
-          {
+          prohibitPath, {
             test: 'data'
           },
           function(e) {
@@ -1099,8 +1053,7 @@ describe(
               .upsertGroup(addedTestGroup)
               .then(function() {
                 testClient.set(
-                  prohibitPath,
-                  {
+                  prohibitPath, {
                     test: 'data'
                   },
                   function(e) {
@@ -1118,19 +1071,17 @@ describe(
         testClient.onSystemMessage(function(eventType) {
           if (eventType === 'server-side-disconnect') {
             testClient.set(
-              '/TEST/a7_eventemitter_security_access/' + test_id + '/set',
-              {
+              '/TEST/a7_eventemitter_security_access/' + test_id + '/set', {
                 test: 'data'
-              },
-              {},
+              }, {},
               function(e) {
                 if (!e) return done(new Error('this should not have been allowed...'));
                 expect(e.toString()).to.be('Error: client is disconnected');
 
                 adminClient.set(
-                  '/TEST/a7_eventemitter_security_access/' + test_id + '/set',
-                  { test: 'data' },
-                  {
+                  '/TEST/a7_eventemitter_security_access/' + test_id + '/set', {
+                    test: 'data'
+                  }, {
                     onBehalfOf: testClient.session.user.username
                   },
                   function(e) {
@@ -1142,9 +1093,45 @@ describe(
             );
           }
         });
-
         serviceInstance.services.security.users.deleteUser(addedTestuser, function(e) {
           if (e) return done(e);
+        });
+      });
+
+      xit(`deletes the test1 user, tests we are notified about the session closure,
+          then have no access, we retain the token, recreate the test1 user and
+          we check we cannot reuse the retained token`, function(done) {
+
+        let retainedToken = testClient1.session.token;
+        let testPath = '/TEST/a7_eventemitter_security_access/' + test_id + '/web_access';
+
+        testClient1.onSystemMessage(function(eventType) {
+          if (eventType === 'server-side-disconnect') {
+            testClient1.set(
+              testPath, {
+                test: 'data'
+              }, {},
+              async (e) => {
+                if (!e) return done(new Error('this should not have been allowed...'));
+                expect(e.toString()).to.be('Error: client is disconnected');
+                let firstResponse = await doRequest(testPath, retainedToken);
+                console.log('firstResponse', firstResponse.statusCode);
+                expect(firstResponse.statusCode).to.be(403);
+                await recreateAddedTestUser();
+                let secondResponse = await doRequest(testPath, retainedToken);
+                console.log('secondResponse', secondResponse.statusCode);
+                expect(secondResponse.statusCode).to.be(403);
+                done();
+              }
+            );
+          }
+        });
+        doRequest(testPath, retainedToken).then((controlResponse) => {
+          console.log('controlResponse', controlResponse.statusCode);
+          expect(controlResponse.statusCode).to.be(404);
+          serviceInstance.services.security.users.deleteUser(addedTestuser1, function(e) {
+            if (e) return done(e);
+          });
         });
       });
     });